pyproject.toml
[tool.poetry]
name = "oncaller"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
pyyaml = "^6.0.1"
email-validator = "^2.0.0.post2"

[tool.poetry.scripts]
oncaller = "oncaller.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
